mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Example of deploying an MCP server on Cloud Run"
requires-python = ">=3.10"
dependencies = [
    "fastmcp==2.8.0",
    "psycopg2-binary==2.9.10",
]
